Backported most of the TTC implementation of EMF queries

+++ hu.bme.mit.trainbenchmark.generator.rdf/src/main/java/hu/bme/mit/trainbenchmark/generator/rdf/RDFGenerator.java
@@ -80,7 +80,7 @@ public class RDFGenerator extends Generator {
 	}
 
 	@Override
-	protected Object createVertex(final long id, final String type, final Map<String, Object> attributes,
+	protected Object createVertex(final int id, final String type, final Map<String, Object> attributes,
 			final Map<String, Object> outgoingEdges, final Map<String, Object> incomingEdges) throws IOException {
 
 		// vertex id and type
